pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pet-train"
version = "0.1.0"
requires-python = ">=3.11,<3.12"

dependencies = [
    "pet-schema",
    "torch>=2.1",
    "torchaudio>=2.1",
    "pytorch-lightning>=2.1",
    "wandb",
    "pyyaml",
    "tenacity",
    "python-json-logger",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["pet_train", "pet_train.*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
